Workaround for late_start not firing.
This probably needs a real fix. I guess that mounting /data fails in the android container, and because the partition is marked as being "encryptable", android assumes that it is encrypted and mounts a tmpfs for /data. Then vold.decrypt is set to 1 to indicate that the GUI should ask for a passphrase, which Ubuntu Touch doesn't support, and /data is never mounted correctly, so late_start never gets started.
So the actual fix for this is to fix the /data mounting. I don't really understand the lxc container magic, so I have no idea why it fails..