Hack-fix adb in recovery.
Writing 1 to /sys/class/android_usb/android0/enable some time after booting
makes adb work in recovery. This is supposed to be done in the recovery
init.rc, but it doesn't work when doing it right after boot for whatever
reason. Some delay is required. Add that delay in a custom
init.recovery.vision.rc file.