Fix custom boot.img creation
While copy-pasting stuff from the main Makefile I forgot two important
lines, causing the custom boot.img to be broken. I never noticed it
because there are two different boot.imgs being built: The standalone
boot.img and one that is zipped into the OTA package. So far I've only
used the OTA package, but the standalone boot.img was invalid and fails
to boot because the --base argument was missing.